+++ datasets.h
@@ -7,8 +7,8 @@
 
 typedef struct
 {
-    float *x;
-    float *y;
+    double *x;
+    double *y;
     size_t n_samples;
     size_t dim;
 } dataset;
@@ -17,9 +17,9 @@ void free_dataset(dataset *ds);
 
 dataset *load_weather();
  
-float *normalize_mnist_images(unsigned char *images, size_t image_size, 
+double *normalize_mnist_images(unsigned char *images, size_t image_size, 
         size_t n_samples, int for_gan);
 
-float *one_hot(unsigned char *labels, size_t n_classes, size_t n_samples);
+double *one_hot(unsigned char *labels, size_t n_classes, size_t n_samples);
 
 dataset *load_mnist(int for_gan);
