+++ linalg.h
@@ -9,21 +9,21 @@
 #include <stddef.h>
 
 
-float *vector_addition(float *v1, float *v2, const size_t dim);
-float *vector_subtraction(float *v1, float *v2, const size_t dim);
-float vector_sum(float *v, const size_t dim);
-float *vector_scalar_product(float *v, float s, const size_t dim);
-float *vector_scalar_addition(float *v, float s, const size_t dim);
-float *vector_elementwise_product(float *v1, float *v2, const size_t dim);
-float vector_dot_product(float *v1, float *v2, const size_t dim);
-float **vector_outer_product(float *col, float *row, const size_t col_dim,
+double *vector_addition(double *v1, double *v2, const size_t dim);
+double *vector_subtraction(double *v1, double *v2, const size_t dim);
+double vector_sum(double *v, const size_t dim);
+double *vector_scalar_product(double *v, double s, const size_t dim);
+double *vector_scalar_addition(double *v, double s, const size_t dim);
+double *vector_elementwise_product(double *v1, double *v2, const size_t dim);
+double vector_dot_product(double *v1, double *v2, const size_t dim);
+double **vector_outer_product(double *col, double *row, const size_t col_dim,
         const size_t row_dim);
-float *matrix_vector_dot_product(float **m, float *v, const size_t n_rows, 
+double *matrix_vector_dot_product(double **m, double *v, const size_t n_rows, 
         const size_t n_cols);
-float **matmul(float **m1, float **m2, const size_t n_rows_m1,
+double **matmul(double **m1, double **m2, const size_t n_rows_m1,
         const size_t n_cols_m1, const size_t n_cols_m2);
-float **matrix_transpose(float **m, const size_t n_rows, const size_t n_cols);
-float **matrix_scalar_product(float **m, float s, const size_t n_rows,
+double **matrix_transpose(double **m, const size_t n_rows, const size_t n_cols);
+double **matrix_scalar_product(double **m, double s, const size_t n_rows,
         const size_t n_cols);
-float **matrix_scalar_addition(float **m, float s, const size_t n_rows,
+double **matrix_scalar_addition(double **m, double s, const size_t n_rows,
         const size_t n_cols);
