+++ random.h
@@ -2,5 +2,5 @@
 #include <stddef.h>
 
 
-float random_normal();
-float glorot_random_normal(size_t in_dim, size_t out_dim);
+double random_normal();
+double glorot_random_normal(size_t in_dim, size_t out_dim);
